Implemented changes from the review
